Clean up more
