+++ packages/gamut-system/src/styleTemplates/createResponsiveStyleTemplate/__tests__/responsiveProperty-test.ts
@@ -1,7 +1,7 @@
 import { createResponsiveStyleTemplate } from '..';
 import { DEFAULT_MEDIA_QUERIES } from '../constants';
 import { createStandardStyleTemplate } from '../../createStandardStyleTemplate';
-import { ResponsiveProp } from '../../../types/system';
+import { ResponsiveProp } from '../../../types/config';
 
 describe(createResponsiveStyleTemplate, () => {
   const styleTemplates = {
